requirements.txt
asgiref==3.8.1
Django==5.0.6
pillow==10.4.0
psycopg2==2.9.9
python-decouple==3.8
sqlparse==0.5.0
tzdata==2024.1


